*draft* [display] Allow lower brightness setting
Currently display brightness can be set to values 1-5, which corresponds
to 20%, 40%, ..., 100% of the real maximum display brightness. In dark
environments the 20% minimum proves to be still too bright.

This commit expands the mapping so it covers nearly whole brightness
range as supported by the HW, limiting the minimum to 1%, i.e., the
values 1-5 are mapped to 1%, 25%, 50%, 75% and 100% of the real maximum
brightness.